--- GameServerLib/Maps/SummonersRift.cs
@@ -1,579 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Numerics;
-using GameServerCore.Domain;
-using GameServerCore.Domain.GameObjects;
-using GameServerCore.Enums;
-using GameServerCore.Maps;
-using LeagueSandbox.GameServer;
-using LeagueSandbox.GameServer.GameObjects;
-using LeagueSandbox.GameServer.GameObjects.AttackableUnits.AI;
-using LeagueSandbox.GameServer.GameObjects.Other;
-using LeagueSandbox.GameServer.Maps;
-
-namespace MapScripts
-{
-    public class Map1 : IMapScript
-    {
-        public bool HasTopLane { get; set; } = true;
-        public bool HasMidLane { get; set; } = true;
-        public bool HasBotLane { get; set; } = true;
-
-        private static readonly List<Vector2> BlueTopWaypoints = new List<Vector2>
-        {
-            new Vector2(917.0f, 1725.0f),
-            new Vector2(1170.0f, 4041.0f),
-            new Vector2(861.0f, 6459.0f),
-            new Vector2(880.0f, 10180.0f),
-            new Vector2(1268.0f, 11675.0f),
-            new Vector2(2806.0f, 13075.0f),
-            new Vector2(3907.0f, 13243.0f),
-            new Vector2(7550.0f, 13407.0f),
-            new Vector2(10244.0f, 13238.0f),
-            new Vector2(10947.0f, 13135.0f),
-            new Vector2(12511.0f, 12776.0f)
-        };
-        private static readonly List<Vector2> BlueBotWaypoints = new List<Vector2>
-        {
-            new Vector2(1487.0f, 1302.0f),
-            new Vector2(3789.0f, 1346.0f),
-            new Vector2(6430.0f, 1005.0f),
-            new Vector2(10995.0f, 1234.0f),
-            new Vector2(12841.0f, 3051.0f),
-            new Vector2(13148.0f, 4202.0f),
-            new Vector2(13249.0f, 7884.0f),
-            new Vector2(12886.0f, 10356.0f),
-            new Vector2(12511.0f, 12776.0f)
-        };
-        private static readonly List<Vector2> BlueMidWaypoints = new List<Vector2>
-        {
-            new Vector2(1418.0f, 1686.0f),
-            new Vector2(2997.0f, 2781.0f),
-            new Vector2(4472.0f, 4727.0f),
-            new Vector2(8375.0f, 8366.0f),
-            new Vector2(10948.0f, 10821.0f),
-            new Vector2(12511.0f, 12776.0f)
-        };
-        private static readonly List<Vector2> RedTopWaypoints = new List<Vector2>
-        {
-            new Vector2(12451.0f, 13217.0f),
-            new Vector2(10947.0f, 13135.0f),
-            new Vector2(10244.0f, 13238.0f),
-            new Vector2(7550.0f, 13407.0f),
-            new Vector2(3907.0f, 13243.0f),
-            new Vector2(2806.0f, 13075.0f),
-            new Vector2(1268.0f, 11675.0f),
-            new Vector2(880.0f, 10180.0f),
-            new Vector2(861.0f, 6459.0f),
-            new Vector2(1170.0f, 4041.0f),
-            new Vector2(1418.0f, 1686.0f)
-        };
-        private static readonly List<Vector2> RedBotWaypoints = new List<Vector2>
-        {
-            new Vector2(13062.0f, 12760.0f),
-            new Vector2(12886.0f, 10356.0f),
-            new Vector2(13249.0f, 7884.0f),
-            new Vector2(13148.0f, 4202.0f),
-            new Vector2(12841.0f, 3051.0f),
-            new Vector2(10995.0f, 1234.0f),
-            new Vector2(6430.0f, 1005.0f),
-            new Vector2(3789.0f, 1346.0f),
-            new Vector2(1418.0f, 1686.0f)
-        };
-        private static readonly List<Vector2> RedMidWaypoints = new List<Vector2>
-        {
-            new Vector2(12511.0f, 12776.0f),
-            new Vector2(10948.0f, 10821.0f),
-            new Vector2(8375.0f, 8366.0f),
-            new Vector2(4472.0f, 4727.0f),
-            new Vector2(2997.0f, 2781.0f),
-            new Vector2(1418.0f, 1686.0f)
-        };
-
-        private static readonly List<MinionSpawnType> RegularMinionWave = new List<MinionSpawnType>
-        {
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_CASTER,
-            MinionSpawnType.MINION_TYPE_CASTER,
-            MinionSpawnType.MINION_TYPE_CASTER
-        };
-        private static readonly List<MinionSpawnType> CannonMinionWave = new List<MinionSpawnType>
-        {
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_CANNON,
-            MinionSpawnType.MINION_TYPE_CASTER,
-            MinionSpawnType.MINION_TYPE_CASTER,
-            MinionSpawnType.MINION_TYPE_CASTER
-        };
-        private static readonly List<MinionSpawnType> SuperMinionWave = new List<MinionSpawnType>
-        {
-            MinionSpawnType.MINION_TYPE_SUPER,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_CASTER,
-            MinionSpawnType.MINION_TYPE_CASTER,
-            MinionSpawnType.MINION_TYPE_CASTER
-        };
-        private static readonly List<MinionSpawnType> DoubleSuperMinionWave = new List<MinionSpawnType>
-        {
-            MinionSpawnType.MINION_TYPE_SUPER,
-            MinionSpawnType.MINION_TYPE_SUPER,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_MELEE,
-            MinionSpawnType.MINION_TYPE_CASTER,
-            MinionSpawnType.MINION_TYPE_CASTER,
-            MinionSpawnType.MINION_TYPE_CASTER
-        };
-        //General Map variable
-        private IMap _map;
-
-        //Stuff about minions
-        public bool SpawnEnabled { get; set; }
-        private int _cannonMinionCount;
-        private int _minionNumber;
-        private readonly long _firstSpawnTime = 10 * 1000;
-        private long _nextSpawnTime = 10 * 1000;
-        private readonly long _spawnInterval = 30 * 1000;
-
-        //General things that will affect players globaly, such as default gold per-second, Starting gold....
-        public float GoldPerSecond { get; set; } = 1.9f;
-        public float StartingGold { get; set; } = 475.0f;
-        public bool HasFirstBloodHappened { get; set; } = false;
-        public bool IsKillGoldRewardReductionActive { get; set; } = true;
-        public int BluePillId { get; set; } = 2001;
-        public long FirstGoldTime { get; set; } = 90 * 1000;
-        private readonly Dictionary<TeamId, SurrenderHandler> _surrenders = new Dictionary<TeamId, SurrenderHandler>();
-
-        //Some Map-Specific stuff, such as tower models, and Tower Items
-        public TurretType GetTurretType(int trueIndex, LaneID lane)
-        {
-            TurretType returnType = TurretType.FOUNTAIN_TURRET;
-
-            if (lane == LaneID.MIDDLE)
-            {
-                if (trueIndex < 3)
-                {
-                    returnType = TurretType.NEXUS_TURRET;
-                    return returnType;
-                }
-
-                trueIndex -= 2;
-            }
-
-            switch (trueIndex)
-            {
-                case 1:
-                case 4:
-                case 5:
-                    returnType = TurretType.INHIBITOR_TURRET;
-                    break;
-                case 2:
-                    returnType = TurretType.INNER_TURRET;
-                    break;
-                case 3:
-                    returnType = TurretType.OUTER_TURRET;
-                    break;
-            }
-
-            return returnType;
-        }
-        public Dictionary<TeamId, Dictionary<TurretType, string>> TowerModels { get; set; } = new Dictionary<TeamId, Dictionary<TurretType, string>>
-        {
-            {TeamId.TEAM_BLUE, new Dictionary<TurretType, string>
-            {
-                {TurretType.FOUNTAIN_TURRET, "TurretShrine" },
-                {TurretType.NEXUS_TURRET, "TurretAngel" },
-                {TurretType.INHIBITOR_TURRET, "TurretDragon" },
-                {TurretType.INNER_TURRET, "TurretNormal2" },
-                {TurretType.OUTER_TURRET, "TurretNormal" },
-            } },
-            {TeamId.TEAM_PURPLE, new Dictionary<TurretType, string>
-            {
-                {TurretType.FOUNTAIN_TURRET, "TurretShrine" },
-                {TurretType.NEXUS_TURRET, "TurretNormal" },
-                {TurretType.INHIBITOR_TURRET, "TurretGiant" },
-                {TurretType.INNER_TURRET, "TurretWorm2" },
-                {TurretType.OUTER_TURRET, "TurretWorm" },
-            } }
-        };
-
-        public Dictionary<TeamId, Dictionary<MinionSpawnType, string>> MinionModels { get; set; } = new Dictionary<TeamId, Dictionary<MinionSpawnType, string>>
-        {
-            {TeamId.TEAM_BLUE, new Dictionary<MinionSpawnType, string>{
-                {MinionSpawnType.MINION_TYPE_MELEE, "Blue_Minion_Basic"},
-                {MinionSpawnType.MINION_TYPE_CASTER, "Blue_Minion_Wizard"},
-                {MinionSpawnType.MINION_TYPE_CANNON, "Blue_Minion_MechCannon"},
-                {MinionSpawnType.MINION_TYPE_SUPER, "Blue_Minion_MechMelee"}
-            }},
-            {TeamId.TEAM_PURPLE, new Dictionary<MinionSpawnType, string>{
-                {MinionSpawnType.MINION_TYPE_MELEE, "Red_Minion_Basic"},
-                {MinionSpawnType.MINION_TYPE_CASTER, "Red_Minion_Wizard"},
-                {MinionSpawnType.MINION_TYPE_CANNON, "Red_Minion_MechCannon"},
-                {MinionSpawnType.MINION_TYPE_SUPER, "Red_Minion_MechMelee"}
-            }}
-        };
-
-        public Dictionary<TurretType, int[]> TurretItems { get; set; } = new Dictionary<TurretType, int[]>
-        {
-            { TurretType.OUTER_TURRET, new[] { 1500, 1501, 1502, 1503 } },
-            { TurretType.INNER_TURRET, new[] { 1500, 1501, 1502, 1503, 1504 } },
-            { TurretType.INHIBITOR_TURRET, new[] { 1501, 1502, 1503, 1505 } },
-            { TurretType.NEXUS_TURRET, new[] { 1501, 1502, 1503, 1505 } }
-        };
-
-        public void Init(IMap map)
-        {
-            _map = map;
-
-            SpawnEnabled = map.IsMinionSpawnEnabled();
-            map.AddSurrender(1200000.0f, 300000.0f, 30.0f);
-
-            //Due to riot's questionable map-naming scheme some towers are missplaced into other lanes during outomated setup, so we have to manually fix them.
-            map.ChangeTowerOnMapList("Turret_T1_C_06_A", TeamId.TEAM_BLUE, LaneID.MIDDLE, LaneID.TOP);
-            map.ChangeTowerOnMapList("Turret_T1_C_07_A", TeamId.TEAM_BLUE, LaneID.MIDDLE, LaneID.BOTTOM);
-
-            // Announcer events
-            map.AddAnnouncement(30 * 1000, Announces.WELCOME_TO_SR, true); // Welcome to SR
-            map.AddAnnouncement(_firstSpawnTime - 30 * 1000, Announces.THIRY_SECONDS_TO_MINIONS_SPAWN, true); // 30 seconds until minions spawn
-            map.AddAnnouncement(_firstSpawnTime, Announces.MINIONS_HAVE_SPAWNED, false); // Minions have spawned (90 * 1000)
-            map.AddAnnouncement(_firstSpawnTime, Announces.MINIONS_HAVE_SPAWNED2, false); // Minions have spawned [2] (90 * 1000)
-
-            //Map props
-            map.AddObject(new Vector2(12465.0f, 14422.257f), 101.0f, new Vector3(0.0f, 0.0f, 0.0f), 0.0f, 0.0f, "LevelProp_Yonkey", "Yonkey");
-            map.AddObject(new Vector2(-76.0f, 1769.1589f), 94.0f, new Vector3(0.0f, 0.0f, 0.0f), 0.0f, 0.0f, "LevelProp_Yonkey1", "Yonkey");
-            map.AddObject(new Vector2(13374.17f, 14245.673f), 194.9741f, new Vector3(224.0f, 33.33f, 0.0f), 0.0f, -44.44f, "LevelProp_ShopMale", "ShopMale");
-            map.AddObject(new Vector2(-99.5613f, 855.6632f), 191.4039f, new Vector3(158.0f, 0.0f, 0.0f), 0.0f, 0.0f, "LevelProp_ShopMale1", "ShopMale");
-        }
-
-        public void Update(float diff)
-        {
-            if (_map.GameTime() >= 120 * 1000)
-            {
-                IsKillGoldRewardReductionActive = false;
-            }
-
-            if (SpawnEnabled)
-            {
-                if (_minionNumber > 0)
-                {
-                    if (_map.GameTime() >= _nextSpawnTime + _minionNumber * 8 * 100)
-                    { // Spawn new wave every 0.8s
-                        if (Spawn())
-                        {
-                            _minionNumber = 0;
-                            _nextSpawnTime = (long)_map.GameTime() + _spawnInterval;
-                        }
-                        else
-                        {
-                            _minionNumber++;
-                        }
-                    }
-                }
-                else if (_map.GameTime() >= _nextSpawnTime)
-                {
-                    Spawn();
-                    _minionNumber++;
-                }
-            }
-
-            foreach (var fountain in _map._fountains.Values)
-            {
-                fountain.Update(diff);
-            }
-
-            foreach (var surrender in _surrenders.Values)
-                surrender.Update(diff);
-        }
-
-        public string GetMinionModel(TeamId team, MinionSpawnType type)
-        {
-            var teamDictionary = new Dictionary<TeamId, string>
-            {
-                {TeamId.TEAM_BLUE, "Blue"},
-                {TeamId.TEAM_PURPLE, "Red"}
-            };
-
-            var typeDictionary = new Dictionary<MinionSpawnType, string>
-            {
-                {MinionSpawnType.MINION_TYPE_MELEE, "Basic"},
-                {MinionSpawnType.MINION_TYPE_CASTER, "Wizard"},
-                {MinionSpawnType.MINION_TYPE_CANNON, "MechCannon"},
-                {MinionSpawnType.MINION_TYPE_SUPER, "MechMelee"}
-            };
-
-            if (!teamDictionary.ContainsKey(team) || !typeDictionary.ContainsKey(type))
-            {
-                return string.Empty;
-            }
-
-            return $"{teamDictionary[team]}_Minion_{typeDictionary[type]}";
-        }
-
-        public float GetGoldFor(IAttackableUnit u)
-        {
-            if (!(u is ILaneMinion m))
-            {
-                if (!(u is IChampion c))
-                {
-                    return 0.0f;
-                }
-
-                var gold = 300.0f; //normal gold for a kill
-                if (c.KillDeathCounter < 5 && c.KillDeathCounter >= 0)
-                {
-                    if (c.KillDeathCounter == 0)
-                    {
-                        return gold;
-                    }
-
-                    for (var i = c.KillDeathCounter; i > 1; --i)
-                    {
-                        gold += gold * 0.165f;
-                    }
-
-                    return gold;
-                }
-
-                if (c.KillDeathCounter >= 5)
-                {
-                    return 500.0f;
-                }
-
-                if (c.KillDeathCounter >= 0)
-                    return 0.0f;
-
-                var firstDeathGold = gold - gold * 0.085f;
-
-                if (c.KillDeathCounter == -1)
-                {
-                    return firstDeathGold;
-                }
-
-                for (var i = c.KillDeathCounter; i < -1; ++i)
-                {
-                    firstDeathGold -= firstDeathGold * 0.2f;
-                }
-
-                if (firstDeathGold < 50)
-                {
-                    firstDeathGold = 50;
-                }
-
-                return firstDeathGold;
-            }
-
-            var dic = new Dictionary<MinionSpawnType, float>
-            {
-                { MinionSpawnType.MINION_TYPE_MELEE, 19.8f + 0.2f * (int)(_map.GameTime() / (90 * 1000)) },
-                { MinionSpawnType.MINION_TYPE_CASTER, 16.8f + 0.2f * (int)(_map.GameTime() / (90 * 1000)) },
-                { MinionSpawnType.MINION_TYPE_CANNON, 40.0f + 0.5f * (int)(_map.GameTime() / (90 * 1000)) },
-                { MinionSpawnType.MINION_TYPE_SUPER, 40.0f + 1.0f * (int)(_map.GameTime() / (180 * 1000)) }
-            };
-
-            if (!dic.ContainsKey(m.MinionSpawnType))
-            {
-                return 0.0f;
-            }
-
-            return dic[m.MinionSpawnType];
-        }
-
-        public float GetExperienceFor(IAttackableUnit u)
-        {
-            if (!(u is ILaneMinion m))
-            {
-                return 0.0f;
-            }
-
-            var dic = new Dictionary<MinionSpawnType, float>
-            {
-                { MinionSpawnType.MINION_TYPE_MELEE, 64.0f },
-                { MinionSpawnType.MINION_TYPE_CASTER, 32.0f },
-                { MinionSpawnType.MINION_TYPE_CANNON, 92.0f },
-                { MinionSpawnType.MINION_TYPE_SUPER, 97.0f }
-            };
-
-            if (!dic.ContainsKey(m.MinionSpawnType))
-            {
-                return 0.0f;
-            }
-
-            return dic[m.MinionSpawnType];
-        }
-
-
-
-        public void SetMinionStats(ILaneMinion m)
-        {
-            // Same for all minions
-            m.Stats.MoveSpeed.BaseValue = 325.0f;
-
-            switch (m.MinionSpawnType)
-            {
-                case MinionSpawnType.MINION_TYPE_MELEE:
-                    m.Stats.CurrentHealth = 475.0f + 20.0f * (int)(_map.GameTime() / (180 * 1000));
-                    m.Stats.HealthPoints.BaseValue = 475.0f + 20.0f * (int)(_map.GameTime() / (180 * 1000));
-                    m.Stats.AttackDamage.BaseValue = 12.0f + 1.0f * (int)(_map.GameTime() / (180 * 1000));
-                    m.Stats.Range.BaseValue = 180.0f;
-                    m.Stats.AttackSpeedFlat = 1.250f;
-                    m.IsMelee = true;
-                    break;
-                case MinionSpawnType.MINION_TYPE_CASTER:
-                    m.Stats.CurrentHealth = 279.0f + 7.5f * (int)(_map.GameTime() / (90 * 1000));
-                    m.Stats.HealthPoints.BaseValue = 279.0f + 7.5f * (int)(_map.GameTime() / (90 * 1000));
-                    m.Stats.AttackDamage.BaseValue = 23.0f + 1.0f * (int)(_map.GameTime() / (90 * 1000));
-                    m.Stats.Range.BaseValue = 600.0f;
-                    m.Stats.AttackSpeedFlat = 0.670f;
-                    break;
-                case MinionSpawnType.MINION_TYPE_CANNON:
-                    m.Stats.CurrentHealth = 700.0f + 27.0f * (int)(_map.GameTime() / (180 * 1000));
-                    m.Stats.HealthPoints.BaseValue = 700.0f + 27.0f * (int)(_map.GameTime() / (180 * 1000));
-                    m.Stats.AttackDamage.BaseValue = 40.0f + 3.0f * (int)(_map.GameTime() / (180 * 1000));
-                    m.Stats.Range.BaseValue = 450.0f;
-                    m.Stats.AttackSpeedFlat = 1.0f;
-                    break;
-                case MinionSpawnType.MINION_TYPE_SUPER:
-                    m.Stats.CurrentHealth = 1500.0f + 200.0f * (int)(_map.GameTime() / (180 * 1000));
-                    m.Stats.HealthPoints.BaseValue = 1500.0f + 200.0f * (int)(_map.GameTime() / (180 * 1000));
-                    m.Stats.AttackDamage.BaseValue = 190.0f + 10.0f * (int)(_map.GameTime() / (180 * 1000));
-                    m.Stats.Range.BaseValue = 170.0f;
-                    m.Stats.AttackSpeedFlat = 0.694f;
-                    m.Stats.Armor.BaseValue = 30.0f;
-                    m.Stats.MagicResist.BaseValue = -30.0f;
-                    m.IsMelee = true;
-                    break;
-            }
-        }
-
-        public bool Spawn()
-        {
-            var barracks = new List<string>();
-            foreach (var barrack in _map.GetSpawnBarracks())
-            {
-                barracks.Add(barrack.Value.Name);
-            }
-
-            var cannonMinionTimestamps = new List<Tuple<long, int>>
-            {
-                new Tuple<long, int>(0, 2),
-                new Tuple<long, int>(20 * 60 * 1000, 1),
-                new Tuple<long, int>(35 * 60 * 1000, 0)
-            };
-
-            var spawnToWaypoints = new Dictionary<string, Tuple<List<Vector2>, uint>>();
-            foreach (var barrack in _map.GetSpawnBarracks())
-            {
-                if (!barrack.Value.Name.StartsWith("__P"))
-                {
-                    continue;
-                }
-
-                List<Vector2> waypoint = new List<Vector2>();
-                TeamId opposed_team = barrack.Value.GetOpposingTeamID();
-                LaneID lane = barrack.Value.GetSpawnBarrackLaneID();
-
-                if (opposed_team == TeamId.TEAM_PURPLE)
-                {
-                    if (lane == LaneID.TOP)
-                    {
-                        waypoint = BlueTopWaypoints;
-                    }
-                    else if (lane == LaneID.MIDDLE)
-                    {
-                        waypoint = BlueMidWaypoints;
-                    }
-                    else if (lane == LaneID.BOTTOM)
-                    {
-                        waypoint = BlueBotWaypoints;
-                    }
-                }
-                else if (opposed_team == TeamId.TEAM_BLUE)
-                {
-                    if (lane == LaneID.TOP)
-                    {
-                        waypoint = RedTopWaypoints;
-                    }
-                    else if (lane == LaneID.MIDDLE)
-                    {
-                        waypoint = RedMidWaypoints;
-                    }
-                    else if (lane == LaneID.BOTTOM)
-                    {
-                        waypoint = RedBotWaypoints;
-                    }
-                }
-                spawnToWaypoints.Add(barrack.Value.Name, Tuple.Create(waypoint, _map._inhibitors[opposed_team][lane][0].NetId));
-            }
-            var cannonMinionCap = 2;
-
-            foreach (var timestamp in cannonMinionTimestamps)
-            {
-                if (_map.GameTime() >= timestamp.Item1)
-                {
-                    cannonMinionCap = timestamp.Item2;
-                }
-            }
-
-            foreach (var barracksName in barracks)
-            {
-                var waypoints = spawnToWaypoints[barracksName].Item1;
-                var inhibitorId = spawnToWaypoints[barracksName].Item2;
-                var inhibitor = _map.GetInhibitorById(inhibitorId);
-                var isInhibitorDead = inhibitor.InhibitorState == InhibitorState.DEAD && !inhibitor.RespawnAnnounced;
-
-                var oppositeTeam = TeamId.TEAM_BLUE;
-                if (inhibitor.Team == TeamId.TEAM_PURPLE)
-                {
-                    oppositeTeam = TeamId.TEAM_PURPLE;
-                }
-
-                var areAllInhibitorsDead = _map.AllInhibitorsDestroyedFromTeam(oppositeTeam) && !inhibitor.RespawnAnnounced;
-
-                var list = RegularMinionWave;
-                if (_cannonMinionCount >= cannonMinionCap)
-                {
-                    list = CannonMinionWave;
-                }
-
-                if (isInhibitorDead)
-                {
-                    list = SuperMinionWave;
-                }
-
-                if (areAllInhibitorsDead)
-                {
-                    list = DoubleSuperMinionWave;
-                }
-
-                _map.SpawnMinion(list, _minionNumber, barracksName, waypoints);
-            }
-
-            if (_minionNumber < 8)
-            {
-                return false;
-            }
-
-            if (_cannonMinionCount >= cannonMinionCap)
-            {
-                _cannonMinionCount = 0;
-            }
-            else
-            {
-                _cannonMinionCount++;
-            }
-            return true;
-        }
-
-        public void HandleSurrender(int userId, IChampion who, bool vote)
-        {
-            if (_surrenders.ContainsKey(who.Team))
-                _surrenders[who.Team].HandleSurrender(userId, who, vote);
-        }
-    }
-}
